Update to enable args based on files rather than projects
